backend/python_service/requirements.txt
fastapi==0.68.1
uvicorn==0.15.0
pymongo==3.12.0
python-dotenv==0.19.0
requests==2.26.0
pydantic==1.8.2
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.5 